Create Right Angle Triangle.py
Given an integer s. Write a program to print the Right angle triangle wall. The length of perpendicular and base is s.  Use single loop and string multiplication.

Example 1:

Input:
s = 4
Output:

Explanation:
Length of perpendicular and base of triangle is 4 .
Your Task:
The input s is provided as a parameter to the function triangleWall. Complete the given code so that a new line is provided at the end.