Create Inverted Right AngleTriangle.pyREADME.md
Given an integer s. Write a program to print the Inverted "Right angle triangle" wall. The length of the perpendicular and base is s.  Use a single loop and string multiplication.

Example 1:

Input:
s = 4
Output:

Explanation:
Length of perpendicular and base of triangle is 4 .
Your Task:
The input s is provided as a parameter to the function invTriangleWall. Complete the given code so that the inverted triangle is printed. You would have to provide a new line at the end.