Create Divisor.py
Given an integer n. Write a program to print all the divisors of n.

Example 1:

Input:
n = 12
Output: 
1 2 3 4 6 12
Explanation:
12 is divisible by 1 2 3 4 6 12.

Your Task:
The input n is provided as a parameter to the function divisor. Complete the given code so that it prints all the divisors in a single line. Don't print a new line as it will be auto appended by the driver code.