Create Check Palindrome.py
Given a string s, you need to check if it is palindrome or not. A palidrome is a string that reads the same from front and back. Ignore the case in this question.

Example 1:

Input:
s = "Hello"
Output: 
False
Explanation: Hello is not equal to olleH
so it's not a palindrome.
Example 2:

Input:
s = "TenEt"
Output:
True
Explanation: TenEt = tEneT as we are
ignoring the case.
Your Task: 
The input s is provided as a parameter to the function isPalin. You need to return True if it is a palindrome, else return False. The driver code will print the output.