Create Find Pattern.py
Given a string s, and a pattern p. You need to find if p exists in s or not and return the starting index of p in s. If p does not exist in s then -1 will be returned.
Here p and s both are case-sensitive.
Example 1:

Input:
s = "Hello"
p = "llo"
Output: 
2
Explanation: llo starts from the second index in Hello.
Example 2:

Input:
s = "World"
p = "Doodle"
Output:
-1
Explanation: Both are different.
Your Task: 
The input s and p are provided as a parameter to the function findPattern. You need to return the starting index of p in s if it exists, else return -1.