Create List Traversal.py
1. List Traversal 
Basic Accuracy: 77.16% Submissions: 35 Points: 1
You are given a list that contains integers. You need to print the elements of the list with a space between them.

Example 1:

Input:
numbers = [54, 43, 2, 1, 5]
Output: 
54 43 2 1 5
Explanation: Just traverse and print the numbers.
Example 2:

Input:
numbers = [324, 5, 2, 2]
Output:
324 5 2 2
Explanation: Just traverse and print the numbers.
Your Task: 
Write the traversal code to print the elements of the list. The list is provided as the parameter to the function listTraversal. Print the elements with a space between the elements and don't give a new line as it is already provided by the driver code.