Create List Traversal Reverse.py
2. List Traversal Reverse 
Easy Accuracy: 75.79% Submissions: 33 Points: 2
You are given a list that contains integers. You need to print the elements of the list with in reverse order with a space between them.

Example 1:

Input:
numbers = [54, 43, 2, 1, 5]
Output: 
5 1 2 43 54
Explanation: Just traverse in reverse and print the numbers.
Example 2:

Input:
numbers = [324, 5, 2, 2]
Output:
2 2 5 324
Explanation: Just traverse in reverse and print the numbers.
Your Task: 
Write the traversal code to print the elements of the list in reverse order. The list is provided as the parameter to the function listTraversalReverse. Print the elements with a space between the elements and don't give a new line as it is already provided by the driver code.