Create Sum The List.py
You are given a list that contains integers. You need to return the sum of the list.

Example 1:

Input:
numbers = [54, 43, 2, 1, 5]
Output: 
105
Explanation: Just sum it 54+43+2+1+5=105.
Example 2:

Input:
numbers = [324, 5, 2, 2]
Output:
333
Explanation: Just sum it 324+5+2+2=333.
Your Task: 
Write the code to return the sum of the list. The list is provided as a parameter to the function listSum. Don't print the output, just return it as it will be printed by the driver code.