Create Length of The List.py
You are given a list that contains integers. You need to return the length of the list.

Example 1:

Input:
numbers = [54, 43, 2, 1, 5]
Output: 
5
Explanation: The list contains 5 elements so its length is 5.
Example 2:

Input:
numbers = [324, 5, 2, 2]
Output:
4
Explanation: The list contains 4 elements so its length is 4.
Your Task: 
Write the code to return the length of the list. The list is provided as a parameter to the function listLength. Don't print the output, just return it as it will be printed by the driver code.