Create  Append To List.py
You are given three inputs a, b, c. You need to create a list and append a, b, c to the list and then return that list.
Example 1:

Input:
a = 1
b = 2
c = 3
Output: 
1 2 3
Explanation: Just create a list and append 1 2 3 to it [1,2,3]. The return [1,2,3] list.
Example 2:

Input:
a = 4
b = 5
c = 6
Output:
4 5 6
Explanation: Just create a list and append 4 5 6 to it [4,5,6]. The return [4,5,6] list.
Your Task: 
Write the code to create a list. The append a, then append b, then append c to the created list. Finally return the list. The driver code will print the returned list.