Create Count Distinct.py
You are given a list numbers that contains integers. You need to count distinct integers in list.

Example 1:

Input:
numbers = [4, 4, 5, 4, 3, 8, 4, 2, 4, 8,
           1, 7]
Output: 
7
Explanation: The distinct list would be
[4, 5, 3, 8, 2, 1, 7]. The length is 7.
Example 2:

Input:
numbers = [2, 3, 4, 5, 1, 2, 14, 6, 8,
           7, 5]
Output:
9
Explanation: The distinct list would be
[2, 3, 4, 5, 1, 14, 6, 8, 7]. The length
is 9.
Your Task:
Write the code in the function countDistinct. The function takes numbers list as an argument and expects the count of distinct numbers as a return value.